backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib==1.7.4
redis==5.2.1
qrcode[pil]==8.0
pillow==11.1.0
openpyxl==3.1.5
python-multipart==0.0.20
pydantic-settings==2.8.1
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
